* Extract CitiesMockDomainClient to own file * Use mock DomainClient for LoadOperation lifecycle tests for improved reliability
